% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.INLAjoint.R
\name{coef.INLAjoint}
\alias{coef.INLAjoint}
\title{Extracts model coefficients from a given model fitted with INLAjoint}
\usage{
\method{coef}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts model coefficients from INLAjoint objects.
}
