% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.INLAjoint.R
\name{family.INLAjoint}
\alias{family.INLAjoint}
\title{Extracts family from a given model fitted with INLAjoint}
\usage{
\method{family}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts family from INLAjoint objects.
}
