% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.INLAjoint.R
\name{fitted.INLAjoint}
\alias{fitted.INLAjoint}
\title{Extracts fitted values from a given model fitted with INLAjoint}
\usage{
\method{fitted}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts fitted values from INLAjoint objects.
Values are associated to a name to keep track of the outcome related to each fitted value.
}
