% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.INLAjoint.R
\name{fixef.INLAjoint}
\alias{fixef.INLAjoint}
\title{Extracts fixed effects values from a given model fitted with INLAjoint}
\usage{
\method{fixef}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts fixed effects values from INLAjoint objects.
}
