% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.INLAjoint.R
\name{formula.INLAjoint}
\alias{formula.INLAjoint}
\title{Extracts formula from a given model fitted with INLAjoint}
\usage{
\method{formula}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}
}
\description{
This function extracts formula from INLAjoint objects.
}
