% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{INLAjoint.ginv}
\alias{INLAjoint.ginv}
\title{Setup ginv}
\usage{
INLAjoint.ginv(x, tol = sqrt(.Machine$double.eps), rankdef = NULL)
}
\arguments{
\item{x}{input}

\item{tol}{tolerance}

\item{rankdef}{rank
output:}
}
\value{
ginv
}
\description{
Setup random walk
input:
}
