% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{INLAjoint.rw}
\alias{INLAjoint.rw}
\title{Setup rw}
\usage{
INLAjoint.rw(n, order = 1L, sparse = TRUE, scale.model = FALSE, cyclic = FALSE)
}
\arguments{
\item{n}{number of knots}

\item{order}{order 1}

\item{sparse}{boolean, sparsity}

\item{scale.model}{boolean, scale}

\item{cyclic}{boolean, cyclic
output:}
}
\value{
random walk 1
}
\description{
Setup random walk
input:
}
