% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{INLAjoint.rw2}
\alias{INLAjoint.rw2}
\title{Setup rw2}
\usage{
INLAjoint.rw2(n, order = 2L, ...)
}
\arguments{
\item{n}{number of knots}

\item{order}{order 2}

\item{...}{additional arguments passed to INLAjoint.rw
output:}
}
\value{
random walk 2
}
\description{
Setup random walk of order 2
input:
}
