% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{INLAjoint.scopy.define}
\alias{INLAjoint.scopy.define}
\title{Setup scopy}
\usage{
INLAjoint.scopy.define(n = 5L)
}
\arguments{
\item{n}{number of knots
output:}
}
\value{
Matrix with weights for scopy
}
\description{
Setup weights for non-linear effects
input:
}
