% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.rerun.R
\name{joint.rerun}
\alias{joint.rerun}
\title{Rerun a model fitted with INLAjoint}
\usage{
joint.rerun(model, ...)
}
\arguments{
\item{model}{an object containing a model fitted with the joint() function.}

\item{...}{Extra arguments.}
}
\value{
An object of class \code{INLAjoint} containing a model fitted with the joint() function.
}
\description{
Reruns inla() for an object of class \code{INLAjoint} returned by the \code{joint}
  function. The rerun starts with posterior values from previous run and can sometimes improve the model fit
  (for very complex models or unstable parameter estimates due to low information in the data)
}
\seealso{
\code{\link{joint}}.
}
