% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.INLAjoint.R
\name{logLik.INLAjoint}
\alias{logLik.INLAjoint}
\title{Extracts log-likelihood value from a given model fitted with INLAjoint}
\usage{
\method{logLik}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts log-likelihood value from INLAjoint objects.
}
