% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Long.R
\docType{data}
\name{Long}
\alias{Long}
\title{Simulated multivariate longitudinal dataset}
\format{
## `Long`
A data frame with 96 rows and 7 columns:
\describe{
  \item{Id}{Individual id}
  \item{time}{Visit times}
  \item{ctsX}{Continuous covariate}
  \item{binX}{Binary covariate}
  \item{Y1, Y2, Y3}{Gaussian Longitudinal outcomes}
  ...
}
}
\usage{
Long
}
\description{
A simulated dataset used to illustrate some models (e.g., example of joint() function in the documentation). The corresponding survival dataset is named 'Surv'.
}
\keyword{datasets}
