% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongMS.R
\docType{data}
\name{LongMS}
\alias{LongMS}
\title{Simulated univariate longitudinal dataset}
\format{
## `LongMS`
A data frame with 96 rows and 7 columns:
\describe{
  \item{id}{Individual id}
  \item{time}{Visit times}
  \item{X}{Continuous covariate}
  \item{y}{Gaussian longitudinal outcome}
  ...
}
}
\usage{
LongMS
}
\description{
A simulated dataset used to illustrate joint models with univariate longitudinal and multi-state survival component (see joint longitudinal and multi-state model in INLAjoint vignette). The corresponding survival dataset is named 'SurvMS'.
}
\keyword{datasets}
