% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.INLAjoint.R
\name{nobs.INLAjoint}
\alias{nobs.INLAjoint}
\title{Extracts number of observations of each composant from a given model fitted with INLAjoint}
\usage{
\method{nobs}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts number of observations of each composant from INLAjoint objects.
}
