% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.INLAjoint.R
\name{plot.INLAjoint}
\alias{plot.INLAjoint}
\title{Plot the output from a multivariate joint model for longitudinal and/or survival data}
\usage{
\method{plot}{INLAjoint}(jres, sdcor = FALSE, ...)
}
\arguments{
\item{jres}{an object with the output of the the \link{joint} function}

\item{sdcor}{logical indicating if the random effects
correlation are to be shown. If FALSE the covariance is shown.}
}
\description{
This function provide plots for the output of a
multivariate joint model for longitudinal and/or survival data.
The output can be stored into an object and manipulated as
a list of ggplot outputs, described bellow.
}
