% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_predictions.R
\name{print.score_predictions}
\alias{print.score_predictions}
\title{Print method for score_predictions objects}
\usage{
\method{print}{score_predictions}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{"score_predictions"}.}

\item{digits}{number of digits to display. Default is 4.}

\item{...}{additional arguments (ignored).}
}
\description{
Print method for score_predictions objects
}
