% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.used.INLAjoint.R, R/priors.used.R
\docType{data}
\name{priors.used}
\alias{priors.used}
\title{Get prior distributions used for an object of class INLAjoint}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
priors.used

priors.used
}
\value{
A list of the prior distribution used for a model of class INLAjoint fitted with joint().

A list of the prior distribution used for a model of class INLAjoint fitted with joint().
}
\description{
Returns the prior distribution for all the parameters of an object of class \code{INLAjoint} returned by the \code{joint}
  function that fits a joint model to multivariate longitudinal and
  time-to-event data.

Returns the prior distribution for all the parameters of an object of class \code{INLAjoint} returned by the \code{joint}
  function that fits a joint model to multivariate longitudinal and
  time-to-event data.
}
\seealso{
\code{\link{joint}}, \code{\link{summary.INLAjoint}}.

\code{\link{joint}}, \code{\link{summary.INLAjoint}}.
}
\keyword{datasets}
