% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.INLAjoint.R
\name{ranef.INLAjoint}
\alias{ranef.INLAjoint}
\title{Extracts random effects values from a given model fitted with INLAjoint}
\usage{
\method{ranef}{INLAjoint}(object, ...)
}
\arguments{
\item{object}{an object that contains a model fitted with INLAjoint.}

\item{...}{Extra arguments.}
}
\description{
This function extracts random effects values from INLAjoint objects.
}
