% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{setup_FE_model}
\alias{setup_FE_model}
\title{Setup fixed effects part for longitudinal marker k}
\usage{
setup_FE_model(formula, dataset, timeVar, k, dataOnly)
}
\arguments{
\item{formula}{with lme4 format (fixed effects and random effects in the same object)}

\item{dataset}{that contains the outcome}

\item{timeVar}{name of time variable}

\item{k}{identifies the longitudinal marker among 1:K markers}

\item{dataOnly}{boolean for internal use, indicates if only preparing data (i.e., not fitting the model)
output:}
}
\value{
colnames(FE) names of the fixed effects (interactions are separated
by ".X." instead of ":" to facilitate their manipulation)

FE values of the fixed effects
}
\description{
Setup fixed effects part for longitudinal marker k  (internal function)
input:
}
