% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{setup_S_model}
\alias{setup_S_model}
\title{Setup survival part for outcome m}
\usage{
setup_S_model(
  formula,
  formLong,
  dataSurv,
  LSurvdat,
  timeVar,
  assoc,
  id,
  m,
  K,
  M,
  NFT,
  corLong,
  dataOnly,
  SurvInfo,
  strata
)
}
\arguments{
\item{formula}{with inla.surv() object as an outcome}

\item{formLong}{formula from the longitudinal part, if any}

\item{dataSurv}{dataset(s) for the survival part}

\item{LSurvdat}{dataset for the longitudinal part converted to survival format (internal, used to get covariates
if missing in the survival dataset when sharing linear predictors including covariates from longitudinal into survival)}

\item{timeVar}{names of the variables that are time-dependent (only linear for now)}

\item{assoc}{association parameters between K longitudinal outcomes and M survival outcomes (list of K vectors of size M)}

\item{id}{name of the variable that gives the individual id}

\item{m}{identifies the outcome among 1:M time-to-event outcomes}

\item{K}{number of longitudinal outcomes}

\item{M}{number of survival outcomes}

\item{NFT}{maximum number of functions of time (fixed value)}

\item{corLong}{boolean that indicates if random effects across longitudinal markers are correlated,
when multiple longitudinal markers are included in the model}

\item{dataOnly}{boolean for internal use, indicates if only preparing data (i.e., not fitting the model)}

\item{SurvInfo}{information about survival submodels for internal use}

\item{strata}{covariate for stratified proportional hazards model
output:}
}
\value{
YS_data includes the values of the survival outcome and covariates associated to this survival part,
        with the association parameters but the provided id are temporary and they will be updated after
        the cox expansion to make them unique and allow for time dependency

YSformF formula for this survival outcome (not including association parameters)
}
\description{
Setup survival part for outcome m (internal function)
input:
}
