% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFonctions.R
\name{setup_Y_model}
\alias{setup_Y_model}
\title{Setup outcome for longitudinal marker}
\usage{
setup_Y_model(formula, dataset, family, k)
}
\arguments{
\item{formula}{with lme4 format (fixed effects and random effects in the same object)}

\item{dataset}{that contains the outcome}

\item{family}{of the outcome (given to check if the distribution matches but the check is not done yet)}

\item{k}{identifies the longitudinal marker among 1:K markers
output:}
}
\value{
YL.name name of the outcome

YL values of the outcome
}
\description{
Setup outcome for longitudinal marker
input:
}
