% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.used.R
\name{priors.used}
\alias{priors.used}
\title{Get prior distributions used for an object of class INLAjoint}
\usage{
priors.used(x)
}
\arguments{
\item{x}{an object containing a model fitted with the joint() function.}
}
\value{
A list of the prior distribution used for a model of class INLAjoint fitted with joint().
}
\description{
Returns the prior distribution for all the parameters of an object of class \code{INLAjoint} returned by the \code{joint}
  function that fits a joint model to multivariate longitudinal and
  time-to-event data.
}
\seealso{
\code{\link{joint}}.
}
